Common JSON wrapper type for response and request.  #102
